refac: rewrite cjsToEsm.cjs into esm module
